requirements.txt
altair
pandas
requests
python-dotenv
google-generativeai
flask